src/requirements.txt
qrcode[pil]>=7.3
google_cloud_storage>=2.2